core: remove the compid of a device
The device should only depend on the sysid. This means that multiple
components are all in the same device.

This fixes the case where a component such as a camera or gimbal would
connect first and then prevent the communication to the autopilot
because the target component ID would be wrong.