doxygen: create docs only for installed headers
It makes more sense to only create docs for the headers in install
directory since these are the actual public facing files that come with
the library.

This commit also adds the make target `install` which just redirects to
`default install`. This allows the target `html` to depend on `install`.