grpc: robustify gRPC and protobuf installation
It is hard to get the gRPC and protobuf dependencies right.
Therefore, we documented the gRPC and protobuf build.

A problem that came up differently in various distributions was the
protobuf version; gRPC currently requires 3.4.0. We opted to
build and install it manually since the chances that the version which
comes with a Linux distribution matches what we need are very small.
The suggested installation location is to `/opt` in order not to mess
with the host system because having a protobuf version different than
what comes with the the distribution is asking for trouble.

The install location of gRPC and protobuf can then be specified using
environment variables.