mission: add delay before camera action
This adds the possibility to add a delay before taking a picture or
starting a video. This is used to wait for a gimbal to reach its
orientation, or a vehicle to slow down.

The current implementation is not optimal because it is based on the
`LOITER_TIME` waypoint instead of the `NAV_DELAY` command which would be
more appropriate but is not supported (yet).