mission: always send item, even if no change
When we implement mission item download we need to have a unique
mapping from mavlink messages to mission items, otherwise we can't
compare uploaded to downloaded mission items.

Previously, we omitted gimbal and speed settings if the previous mission
item had the same setting. However, this would mean that we don't have a
unique mapping when downloading a mission item.
For instance if we have two mission items with speed 5 m/s, we would only
send one mission item and omit the speed setting in the second
mission item. When downloading the mavlink items for this, we get NAN
for the speed of the second mission item because we don't if it was
actually set or omitted when set.  By always setting the speed, we have
a bit more overhead but it makes it explicit and we don't lose
information when doing multiple transfers.

To save mission item "space" it is always possible to set the speed to
NAN which means no mavlink item is sent for it.