core: add enable/disable API for plugins
By providing an enable/disable API for plugins, we can distinguish
between calls that can/need to be made when a plugin is instantiated and
calls that make sense to make once we are conntected to a vehicle.

As an example, it does not make sense for a plugin to try to get a
parameter if no vehicle is conntected. This only produces a lot of
warnings and errors and uses up CPU.

The new API does not need to be implemented by plugins yet. For a
transitional period, runtime warnings are displayed if the methods are
not implemented.