make: always create openssl headers
This uses the build_Android.sh script to call the openssl configure step
without actually re-building all of openssl and curl.

The configure step is needed in order to be able to access the openssl
headers which are symlinked to a `include` folder, and to get the
opensslconf.h headerfile.