mission: retry downloading mission items
In case any of the MISSION_ITEM or MISSION_REQUEST messages get lost on
the way, we need to retry a couple of times.

We do this with two different timeouts, one for the passive case where
the autopilot pulls (requests) mission items from us and an active case
with a lower timeout where we pull mission items.