core: send HITL bit with DO_SET_MODE command
Commander gets confused if this bit is not set correctly, so we need to
add it.
Also, the _armed and _hitl_enabled flags should both be atomic in order
to be thread-safe, also we should only update the flags if a heartbeat
originates from the autopilot but ignore other components such as the
camera or gimbal.