Replace raw pointers => smart pointers (#314)
Smart pointers provide automatic resource management. In other words,
they implement "Resource Acquisition Is Initialialization programming idiom".
The main goal of this idiom is to ensure that resource acquisition occurs
at the same time that the object is initialized, so that all resources
for the object are created and made ready in one line of code.

`std::unique_ptr` is a smart pointer that owns and manages another object
through a pointer and disposes of that object when the unique_ptr goes out of scope.

Ref 1: https://docs.microsoft.com/en-us/cpp/cpp/smart-pointers-modern-cpp
Ref 2: http://en.cppreference.com/w/cpp/memory/unique_ptr