core, plugins: Revert changes for #331
Ideally routing messages to the intended component is a job of
MAVLink-Router, not that of DroneCore. It would increase unnecessary
complexities in DroneCore. Hence, we're leaving this to MAVLink Router.
It depends on how many manufactures on companion board adopt MAVLink
Router.

In absence of MAVLink Router, we may need to routing. We may add it in
future if required.