build_system: update tinyxml2 and gtest
This resolves a corruption on exit error when running the unit test on
Linux. The error happened when the unit test runner exited even without
calling any tests, and it only happened if we linked to tinyxml2. It was
a very peculiar problem that we worked around by using the system
tinyxml2 on Linux. This is now no longer needed and means that we can
have the same version of tinyxml2 for all builds.