core: remove default for IP for TCP connection
We need to remove the default for the TCP connection because this will
lead to ambiguous function calls if add_tcp_connection() without any
arguments is called which could then be either of the two overloaded
methods.

When using add_tcp_connection with two arguments, the IP needs to be
specified, otherwise one could use add_tcp_connection with one argument.