core: fix lockup of action::takeoff_async
This fixes the takeoff async integration test where we locked ourselves
out because the takeoff command actually does two commands in
succession, first a mode change, then the takeoff command. What happened
is that we locked ourselves out because we sent the second command in
the callback of the first command which still had the lock.