follow_me: use sync param setter, remove default
Since we now have an API to set params sync instead of async we can
simplify the way the configuration is applied.

Additionally, it does not really make sense to change the PX4 default
without asking every time the plugin connects. We should rather fetch
the params in the beginning and have them cached for when we need to
change them.