core: use std::pair instead of std::optional
We don't have C++17 (yet) but we can use std::pair in lieu of it.

Instead of having methods with a return argument as well as by-reference
output argument, it is nicer to have one std::pair as the return
argument only. This commit does this for the `make_command_...`.