core: bugfix for clearing message handlers
This is a bugfix for a bug discovered by Jy.wang.
What happened is that when one plugin would want to clear all its
message handlers, the whole mavlink hander table gets cleared and all
other plugins and the core lose their handlers as well.

This bug only manifests itself in practice if you actually destroy a
plugin before others or the core which happens rarely since mostly
everything together is destroyed at the very end.