core: disable fallback to system ID for now
The fallback to system ID causes problems on systems with multiple
components that might answer but do not actually mean that a vehicle is
connected. For instance the remote controller might also send heartbeats
and trigger a discovery event when really the drone is not discovered
(yet).

In the future we might add an option for UUID vs. system ID but for now
this should resolve the issues described.