core: try to prevent command lockup
This is an attempt to fix a lockup issue that we see in the Swift
example app when the camera capture status subscription is used.
Presumably, the lockup has nothing to do with the actual camera capture
status but it just leads to the right timing to trigger something else.

What happens is that `receive_timeout()` is called but then the command
queue turns out to be empty (returns `nullptr`). After that, no more
commands can be processed and we keep getting the warning:
"Command ack not matching our current command: 257".

This makes sense because we stop doing work, so send any commands as
long as `_state` is `State::WAITING` and there is no other timeout that
would reset this.

Therefore, this change should resolve that case by reseting `_state`, so
that we can keep going.

However, this change does not fix or explain why the queue is empty in
the first place and out of sync with `_state`.