core: try to fix queue and state not in sync
This is an attempt to fix the issue described in the previous commit
message. We should remove any timeout handlers before poping work from
the queue to prevent that `receive_timeout()` is called in the moment
when the queue is already empty and doesn't know what to do about it.

Basically, we always try to follow this order to prevent things out of
sync:

1. Set the `_state`
2. Unregister timeout handler
3. Pop from queue.
4. Call user callbacks.

It is still unclear to me how the lockup case exactly happens. I can't
see how `_state` can ever be out of sync with the queue given the
protecting mutexes.