backend: start GRPC server before connecting
This changes the initialization logic to first start the GRPC server and
then try to connect it all.

This is based on the assumption that we can initialize the GRPC server
and the plugin it has using the default/dummy system and we don't have
to have anything connected yet. Once something connects, the system gets
updated and the plugins start to "work".