core: add interface to flag params as changed
This adds an API for plugins to tell SystemImpl that a param
(probably) has changed and other plugins to subscribe to such updates.

This is not a nice solution, we might want a proper param system with
notifications per plugin about what changed in the future. That will
require some more refactoring.