core/backend: back to version tags
We never managed to complete the versioning using a file and git tags
in CI. Therefore, this reverts it all back to versioning using just git
tags.

The only downside of the current implementation is that it won't work
if the source directory is not a git directory, and it won't be updated
unless `make clean` is used.