mission: make mission item comparison less strict
I think we shouldn't compare doubles and floats of mission items with ==
but only compare it to an accuracy that makes sense for the use case. In
case of latitude and longitude we are limited to 1e-7 anyway because the
MAVLink transport happens over int at 1e7.

Included in this change is to make use of the comparison in the
ingration test.