mission: use correct callback for download timeout
This changes the callback used in the case of a timeout while trying to
download the mission items. Before we used the wrong one, so never got
notified. This was not caught by the integration tests because they
generally only test the happy path given we had no hooks yet for a lossy
link where these things happen.