mission: extend activity states to retry correctly
Previously, the mission protocol was not implemented correctly. What
happened was that the activity state did not reflect the different
states while uploading or downloading a mission and therefore if a
timeout occured we didn't know if we have to retry sending a
`MISSION_COUNT` or a `MISSION_REQUEST_INT`; we always defaulted to
retrying a `MISSION_REQUEST_INT` which mostly worked except if we had
dropped messages in the `MISSION_COUNT` step and so we would request
mission item -1 (the initial value of this variable) which is not
something that exists.