mission: retry sending mission_count as well
We previously did not retry to send mission_count. However, to be robust
for the case that the mission count message does not arrive the first
time, it makes sense to retry it as well.

Since we now retry for everything, we don't need different timeout times
anymore which means the timeout handling can be simplified quite a bit.