camera: hack to enable camera without autopilot
This is a workaround to manually enable the camera plugin if a camera
only is connected. It is needed for tests where we test against a
camera (e.g.  camera-manager) without an autopilot. In this case the
current discovery mechanism doesn't work (or doesn't complete in time)
because we never receive a autopilot version. The workaround is just to
check at 2 Hz if a camera is now connected and in that case initialize
everything.