docker: use su-exec, add Jenkins default user
This is an attempt to get the docker images compatible with Jenkins.
We now add a user with ID 1001, so matching the default Jenkins user
which is used if no $LOCAL_USER_ID is supplied.

Also, we use su-exec instead of su hoping because it seems to print less
errors.