mission: fix mission upload cancel crash
This fixes a crash that happened when a mission upload was cancelled.
What happened is that the mission result callback ended up being called
twice on upload after a cancellation.

The reason for the two callback calls are, I believe, as follows:
- The autopilot is still in upload mode. Therefore, it will continue to
  send mission items continuing where we sent cancel. This means we will
  continue happily uploading from there and eventually report SUCCESS.
- However, since the autopilot is still in upload mode, it doesn not
  acknowledge our message indicating the count of the mission items.
  We receive this nack and report an error.

The fix for now is to abort immediately if we receive a nack.

The proper fix is for the mission plugin as well as the PX4 side to
implement mission cancelling using MAV_MISSION_OPERATION_CANCELLED to
avoid busy states like this.