mavlink_parameters: fix timeout for 16 char params
This fixes a char array overflow in strlen in string compare. This
happened because the case where the MAVLink param name has length 16 and
is not handled correctly. The MAVLink field is not 0 terminated which
means that we need to copy to bytes to a 0 terminated buffer first
before using it.