mavlink_passthrough: use thread pool for callbacks
This resolves the case where an API user tries to unregister while still
in the callback. In that case the callback gets erased from the
mavlink_handler_table and this causes the caller's scope to be
destroyed. We can prevent this by copying callback and the message
argument and add it to the user callback queue. Presumably, this can
have performance implications but I think we want to make sure not to
crash first.