backend: only install libbackend.so, not libbackend.a
Installing libbackend.a is annoying because we then need to install
the *.pb.h and *.grpc.pb.h headers, and anyway we don't use it.

So now, when BUILD_SHARED_LIBS=ON, we build and install libbackend.so.
When BUILD_SHARED_LIBS=OFF, we don't install libbackend. However,
backend_bin is always installed, and uses libbackend (static or
shared, depending on BUILD_SHARED_LIBS).