clang-format: AlignAfterOpenBracket -> AlwaysBreak
This changes the chosen clang format style.

The main reason for this change is to work around a clang-format bug
in AlignAfterOpenBracked introduced with clang-format 8:
https://bugs.llvm.org/show_bug.cgi?id=41676

By changing to this option, we resolve the problem were different
versions enforce different styles.

Also, in my opinion this style makes more sense in general.
For some argumentation by Kevlin Henney, I suggest to watch:
https://youtu.be/ZsHMHukIlJY?t=1027