camera: fixes to prevent deadlocks
We have been seeing deadlocks connected with the camera plugin
presumably due to deadlocks which then block the receiving thread.

On inspection I found a couple of callbacks from the camera plugin which
had not been using the thread pool. Hopefully fixing these to all use
the thread pool might fix the issues.