camera: fix deadlock on refresh_calls
It turns out that we wait for our own lock if we do a call to
refresh_params within already doing parameter work. We use an ugly
workaround here of scheduling the refresh on the thread pool. In the
future we should probably use an event loop with something like signals
and slot and get rid of all that nasty locking.

In this change we also make sure to always check the result of
set_setting which we forgot to do before.