add pollin to serial connection read
On an intel nuc running ubuntu I had issues where my sdk example
(mavlink_ftp_client) was returning 1 in the main but the serial connection
thread was blocked because of the blocking read.

Also the close of the file descriptor in the stop method was called before
joining the receive thread which could lead to undefined behavior.