core: fix deadlock on destruction of plugins
This is a fundamental change of ownership of a system. Instead of
letting all the plugins have shared ownership over a system and thus
the mavsdk object, they now only have a raw pointer to the system
(_parent).

I'm not sure if this should be the design going forward but it does fix
weird aborts that we have been seeing on destruction in the integration
tests such as:

terminate called after throwing an instance of 'std::system_error'
  what():  Resource deadlock avoided

What happened is that the plugin was owning the system and mavsdk
through a shared_ptr. For some constellations with callbacks this meant
that a plugin was the last one to go out of scope and therefore trigger
the destruction of the system and mavsdk. If this happened out of a
callback from a plugin it would happen on the work queue which means
that it would try to join and destruct the work queue from the work
queue. Trying to join its own thread would lead to a deadlock and hence
triggers the abort.

Alternatives:
- Using weak_ptr in the plugins but this means the plugin would have to
  keep checking that _parent is still there, and it's not clear what it
  should do without _parent.
- I tried not to join its own thread but that just leads to an abort or
  undefined behavior when it later tries to delete itself.
- I wonder if we could fix that shared_ptr with custom destructors?
- Or we should re-evaluate how all these callbacks work and use some
  aio library and event loop or signals and slots, etc.