mission: flatten plugin, remove MissionItem class
In order to prepare for the auto-generation of the C++ part of the
mission plugin, it probably makes sense to flatten the mission plugin
and make the API less complex. Instead of having this MissionItem class
we now just use a simple struct.

The disadvantage is that we can't check the "settings" of a mission item
at construction but only when it is used as part of a mission upload.

Also, it would be nice to generate the current NAN defaults for floats
in a mission item.