mission: use vector of values instead of pointers
Instead of allocating a shared_ptr for each mission item in the vector,
we now just copy the struct.

I think it was actually a wrong design choice to use shared_ptr for the
mission items because once the items are passed e.g.
upload_mission_async we are passing ownership and we're not supposed to
change these items, otherwise we would mess with them while they are
being uploaded.

In terms of performance a vector of pointers makes sense if reordering
is needed often which is however, not the case for us. So for our case
having the items consecutively in memory is probably faster. (Of course
we won't know until we profile.)