third_party: build absl separate as package
This is needed because otherwise gRPC doesn't install correctly
otherwise.
The underlying issue is that gRPC dependencies don't work as "module"
if the cmake version is lower than 3.13, so they need to be marked as
"package" and added separately, as we already do it for a couple of
other dependencies.