core: move queue from messages to callbacks
Instead of having a queue for incoming messages, it probably makes more
sense to queue the user callbacks instead. This allows MAVSDK internally
to keep working even if the user is blocking in a user callback.

This means that user callbacks are always called from the same thread,
rather than from either the main MAVSDK thread or the receiver thread,
and it also allows for proper warnings if a user does actually block in
a callback.