integration_tests: use thread instead of async
It turns out std::async was not the right choice for this because it
actually returns a future. We then need to wait for the future to
trigger the deferred lambda to run, otherwise it will never actually
run. This means that the actual callback is still held up and we can't
return.

The better approach therefore is to do the work in a thread and detach
it.