core: move timeout check, user callback to mavsdk
I think it makes more sense to have the callback queue thread as well as
the timeout check thread in MavsdkImpl instead of SystemImpl because it
is not actually System specific, so when dealing with multiple systems
this would even safe some threads.

However, and most importantaly, this fixes an abort on destruction
happening in CI integration tests fairly often. What happens is that
sometimes a plugin which is instantiated in a shared_ptr is last used in
a user callback. This means that the destructor of the plugin as well as
SystemImpl is triggered inside a user callback. This leads to an abort
when SystemImpl tries to join the user callback thread, so itself.