mission: add support for gimbal v2 command
This adds support to inject the gimbal attitude using the MAVLink gimbal
v2 protocol instead of the deprecated v1 protocol.

The v1 protocol is still useable if the autopilot does not seem to
support v2, so if it does not respond to the request for
GIMBAL_MANAGER_INFORMATION.

The auto-detection is, for now, duplicated from the gimbal plugin. The
idea is that plugins don't depend on each other to prevent dependency
problems. In the future we might consider moving the v1 vs v2 check into
the system but then again, we don't want to litter the system too much
either. The hope is that v1 won't be around too much longer and we can
just remove the magic.