core: move sending heartbeats to Mavsdk class
It actually did not make sense to send heartbeats in the System class
because it's the Mavsdk MAVLink node that sends them as broadcast and
not per vehicle.

The change was required in order to enable sending heartbeats without
first discovering a system. That's needed if MAVSDK is used on a
companion or as an autopilot instead of as a ground station.

A few specifics for the change:
- Move call_every_handler from SystemImpl to MavsdkImpl.
- Add config option to always send heartbeats, and not just when a
  vehicle has been discovered.
- Remove duplicate system_id/component_id vars in MavsdkImpl and use
  vars in configuration instead.