core: immediately call new entry
Instead of waiting for one interval we now call a new call every
callback immediately in the first round. This is achieved by setting the
initial time to now and subtract the interval.

This should fix the problem where the takeoff and hover test fails due
to the vehicle going into failsafe. This happened because the vehicle
had not received any heartbeats yet at the time of the takeoff.