core: Add plugin access to system, unify libs
This adds the possibility to get plugins from a system like this:

mavsdk.systems()[0].action().takeoff()

This makes the Python interface based on pybind11 much easier because it
solves the ownership problems that we had because of having Mavsdk and
plugins separately instantiated.

This however means that we will have to unify the libraries, so create
one libmavsdk.so combining the core and the plugins.