core: fix CI connection race issues
Hopefully this will fix the issue where CI fails because the on_change
callback is called, however, the system is not actually connected yet.

I believe the issue we saw was a race between the on_change callback
being registered and the moment when a system is added to the list but
has not been set to connected yet.