mission: move QGC plan import from mission to mission_raw
This moves the method to import QGroundControl .plan JSON mission files
from the mission plugin to the mission_raw plugin.

The reasoning is as follows:
The imported missions did not really fit the Mission API, and I guess
that the mission progress was often wrong. When the missions were
imported into mission, many items/commands were unsupported and dropped
with a warning. Such a warning is not visible in the language wrappers
and could lead to surprises. The documentation said that the method
would fail if any unsupported item was being imported, however, that
was not true anymore in code.

Therefore, in my opinion, it makes a lot more sense to implement the
import in mission_raw where all/any mission items and commands can be
imported. It's should be more complete, without surprises, and
therefore safer to use.

And in later steps we can also add geofence and rally items to this API.

This PR contains the proto change and the implementation in code. It
also slightly improves the parsing by moving it to a separate file and
adding a bit more unit test coverage.