core: add filter for messages from QGC
PX4 is now forwarding messages between the SDK link on UDP port 14540
and the QGC link on UDP port 1550. This causes examples, integration
tests, and existing implementations to break as they discover 2 systems
when they are only expecting one.

Therefore, I suggest to add a workaround to filter messages coming from
QGC for the case where MAVSDK is used in the default configuration, so
as a ground station.

The correct fix will be to properly discover the system by checking if
it `has_autopilot()` but it will take some time for that to take place,
so I suggest this workaround in the meantime.