README: rename develop to main branch
We no longer maintain a develop and a master branch and instead just use
the main branch going forward.

The idea behind the two branches was to be able to backport bugfixes or
security fixes onto master, however, in practice we never did that.

Instead we just move ahead release by release. If someone requires
bugfixes on older releases they need to port these themselves. All pull
requests are labelled which should make it easy to find the bugfixes.