core: use vector for systems
When a new system is added we expect it to be added to the end of the
`systems()` vector. This can easily be accomplished by using a
std::vector for the systems instead of a std::unordered_map.

For a small number of systems using a vector for that sort of lookup
should be faster anyway than a map, according to common advice.