core: fix own sysid/compid in mission transfer
It turns out that SystemImpl did not properly implement the Sender class
and therefore own sysid and compid were always 0.

This refactoring fixes that by using methods instead of variable
references of variables that are not set and these abstract methods then
need to be implemented, otherwise the compiler would complain.

I also had to remove on check in the unit tests, why is a mystery to me.