core: potential fix for mission transfer timeout
This could potentially fix an issue where we saw timeouts when uploading
mission items. Basically, the change increases our timeout waiting for
the autopilot to request the next mission item. So if the autopilot
takes a bit longer, we should not give up too early. The timeout is
retries*timeout instead of the timeout only.