core: fix calling promise twice
If we deny a command because the command ID is already in the queue of
commands being sent, then we need to remove it out of the work queue as
well. Otherwise, we might call the callback yet again when an ack
arrives that matche the command in the queue.