camera: handle empty camera definition URI
Currently, MAVSDK expects all cameras to have a camera definition URI
and the Prepare method of the camera plugin will fail after 3 tries
to fetch the file.

However the spec says this is optional, so an empty string should be
allowed and should avoid the error.

This adds a check for the empty string to should_fetch_camera_definition()
to handle this case.