camera: only request flight info when system has autopilot
Currently, the camera plugin assumes that the system was an autopilot
and always requests flight information. However it may be the case that
the system is not an autopilot, e.g. a stand-alone camera.

This adds a condition to check that the system is an autopilot before
sending the request.