core: refactor where reserved value is applied
Instead of setting the reserved value of NAN or 0 in the constructor of
a command, we now do it later when sending it. Until then we keep track
of the params using std::optional. By setting the reserved values later
like this, we can avoid the dependency to SystemImpl in the
CommandLong/CommandInt constructor which makes it nicer to use.