telemetry: use SYS_STATUS to determine calibration
PX4 by now sends the calibration state out using SYS_STATUS. This is to
be preferred as it updates after a calibration unlike the params which
are only fetched on startup.

Hopefully in the future, we can deprecate support for older PX4 versions
and get rid of these params completely.

This commit also refactors the way the ArduPilot params are stored a
bit.