core: requesting a message needs a target
I don't think it is really possible to use the REQUEST_MESSAGE command
without a target_component set. If we were to send out REQUEST_MESSAGE
with target_component set to 0 (broadcast) it would open up a race
between various components that do support it or don't support it. All
of these would respond and it would be tricky to know how to handle all
the acks.

Therefore, I decided to send out these REQUEST_MESSAGE commands with a
target instead. This means that in component_information we will
ultimately have to go through all connected components one by one to
figure out what is there.