core: fix MAV_TYPE of mavsdk_server
It turns out that we sent out MAV_TYPE_GENERIC for UsageType::Custom if
we manually set sysid and compid. However, when we use the usual ground
station compid, we would probably expect that the normal MAV_TYPE_GCS.

This fixes mavsdk_server were sysid and compid is set manually and
therefore the MAV_TYPE was suddenly wrong.