core: param refactor
This changes the param class quite a bit. This is a step towards better
compatibility with the MAVLink spec, specifically support of int types
other than int32_t, e.g. uint8_t.

Since the MAVSDK API only offers int and float params, we need to handle
these other int types internally. This is done by looking up the param
in the cache if available, and if not a param_get is done before doing a
param set.

With these changes we should now also correctly translate float to int
and int to float for the two implementations, so bytewise copying for
PX4 and casting to float for ArduPilot.