core: workaround to avoid deadlocks
We add the start_sending_heartbeats() call the thread queue to avoid
deadlocks.

What happens is that the mavlink receiver mutex is locked when a
heartbeat arrives which then creates the server component to send out
heartbeats. Creating the server component involves registering for
mavlink messages which is what blocks.