params: fix double promise call
This fixes a segfault/abort when using get_all_params(). The issue was
that the timeout handler was not removed as it should have been, and so
after a successful completion, we got another call from the timeout
handler which then satisfies the promise a second time which is not
possible.

This did not show up in the integration test (or at least not often)
because the test was already destructed by the time the timeout would
trigger.