core: second in flight command fix
In order to prevent too many commands from clogging the queue, we remove
them if they look identical.

This changes it so that we only drop a command if it does not come with
a callback. If it has a callback associated with it, it's better to call
it anyway, otherwise odd races or just Denied results when you don't
expect it tend to be the result.

This is the same as I fixed previously, however, this time also for
CommandLong which I forgot last time.