core: improve subscription race
Often when waiting for a system, you only really want one, so the first
one connecting. Therefore, you might immediately unsubscribe again
inside the callback.

This pattern didn't work anymore with the new returned handle because
we did not return from the function before calling the callback meaning
the handle is not actually valid yet.

By adding the callback to the queue we can work around this case
although the race condition still exists until we change when teh
callback is called in a later commit.

Signed-off-by: Julian Oes <julian@oes.ch>