mavsdk_server: simplify linking
On Windows, the mavsdk_server unit tests don't link properly when
building with shared libs. For whatever reason the transitive
dependencies didn't get tracked from the protobu/grpc components all the
way into the mavsdk_server unit tests.

I suspect that either some symbols get stripped too early, or it has to
do with the export visibility in MSVC.

I couldn't quite get to the bottom of it but I realized that we could
just simplify it by avoiding the little libraries and just adding all
sources to the mavsdk_server library in the first place.

Let's see if it works...

Signed-off-by: Julian Oes <julian@oes.ch>