rtk: parse Python RTCM strings
This adds a hack to parse incoming RTCM string from Python. Since they
are somewhat escaped we need to translate them back to the original
bytes.

We do this in v1.4 in order not to break the API. In v2, we hopefully
can switch to a vector<uint8_t> instead.

Signed-off-by: Julian Oes <julian@oes.ch>