core: extract autopilot, move sender
This is a refactor to move the Sender interface to MavsdkImpl as that's
where it fits better than in SystemImpl. That's because a sender has to
do with the origin (which is MavsdkImpl or ComponentImpl) rather than
the target representation (which is SystemImpl).

While doing that we also pulled out the Autopilot type, and cleaned up
the target_system_id variable a bit.

Signed-off-by: Julian Oes <julian@oes.ch>