core: fix burst implementation
This fixes the burst implementation which was implemented according to
the docs rather than the existing implementations out there (like
ArduPilot, MissionPlanner, and QGroundControl.

The difference is that the size field is used to signal the size of each
burst packet rather than the overall burst. This means a burst transfer
always has to be the whole file and can't be a partial file.

Signed-off-by: Julian Oes <julian@oes.ch>