Be able to use mavsdk library in debug and release mode side-by-side.
Since it's not possible to mix debug and release msvc runtime on
windows, it's important to link against a mavsdk debug lib when
compiling the program for debug and later link against the mavsdk
release lib when deploying the program. But this did not work until now
because the name for the debug and release dll was the same. For this
purpose CMake invented the 'DEBUG_POSTFIX' property for a target.
The generated CMake files to find mavsdk are correctly filled with the
appropriate library name so no need to change anything else here.