core: require configuration for Mavsdk ctor
We no longer allow Mavsdk to be instantiated by default. Instead a
configuration with a component type of the Mavsdk component needs to be
supplied.

The reasoning behind the (breaking) change is that we no longer want to
assume a potentially safety critical/dangerous default.

In the past Mavsdk would identify itself as a ground station, as that
likely is the most used purpose. However, when used as a companion
computer, this can likely be dangerous if the component type is not
changed to companion computer. It means that the autopilot will assume
that it has connection with a ground station even though it does not but
only to a local companion computer process.

With this change, the component type needs to be passed to the
construction of Mavsdk, so it's an explicit choice whenever it is used.

Signed-off-by: Julian Oes <julian@oes.ch>