docker: remove unused SITL testing images
We no longer maintain and use the integration tests and SITL tests.

To do this right, we would need a lot more testing infrastructure, and
would have to keep up to date with all PX4 and ArduPilot releases.

I don't have bandwidth to do that, and I don't think it should be up to
MAVSDK to keep a docker image to run PX4 and ArduPilot.

Hence, I'm removing it all.

Signed-off-by: Julian Oes <julian@oes.ch>