docker: update tooling to use clang-format-18
With Ubuntu 24.04 out, we can switch to clang-format-18.

Also, while doing that, I took the liberty to purge the mavsdk-dev
docker image, and replace it with a specific image just for
clang-format. This is all I ever use it for and I don't think a dev
docker image makes much sense or is really needed. It just makes the
formatting image bloated.

Also, I've cleaned up the docker/podman command detection, as well as
made the docker.io repository explicit.