mavsdk_server: fix autopilot discovery
Instead of using the subscribe_on_new_systems callback, we just use a
while loop here. That's because the subscribe_on_new_system callback
won't tell us if an autopilot is discovered when another component of
the same system is discovered first, thus triggering the callback early
when no autopilot is around.

With a stupid old while loop we can just keep checking this until we
have an autopilot and then exit.