core: prevent invalid duplicate acks
It turns out we respond with two acks to a MAV_REQUEST_MESSAGE when we
should only respond once with the successful one.

Alternatively, we could first grab all the results and then do some
magic to determine which one is the best ack but that's a bit too magic
and complex, so I'm not doing that yet.