core: remove shutdown magic, use timeout instead
Shutting down and/or closing the socket from a different thread to stop
the receiver thread lead to a data race on the socket (or socket fd).

Instead it's cleaner to use a timeout to have the receive thread check
the exit flag every 500ms and if it's set exit, so it can be joined and
sockets can be closed properly.

This was found using thread sanitizer.