CI: remove hunter support
I don't actually like if dependency management has to be embedded within
the build. Imagine if we had to put conan, vcpkg, hunter, brew, pacman,
etc. you name it, all within cmake. It would be a mess.

We strive for just simple plain cmake instead.

The reason this is removed now is that I need to add the dependency
picosha2 and I can't find it in hunter. Also, I haven't seen any signs
that hunter support is actually (still) used out there in the wild.

For now, let's remove it. If this is still needed, someone will complain
and help add it back in.