core: fix regression in locked_queue
It turns out that I broke the LockGuard recently by adding a unique_lock
which is of course unlocked when it goes out of scope, when it should be
stayed locked.

This fix instead uses a unique_lock as a member variable rather than
manually locking and unlocking in the constructor and destructor.